Refactor AssemblyExtensions and update README format
AssemblyExtensions has been significantly refactored with major changes: the class was renamed to HotLoader, and has been turned into an instance class with reduced methods. The README file format was updated for improved readability by center aligning the title and description. A reference to the Examples project was also added to the solution file.