Update README and remove dependency on ImmediateReflection
The README was updated to reflect changes in how the software is used. The dependency on ImmediateReflection was also removed, which involved changes to the project file, and the corresponding download and license information was removed from the build target. The AssemblyExtensions file was also updated to reflect this change.